% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_set_ID}
\alias{add_set_ID}
\title{Helper function for generating lists of equipment with corresponding IDs and
column names}
\usage{
add_set_ID(df, ID, col_name)
}
\arguments{
\item{df}{data frame of item numbers and massor other property}

\item{ID}{the character string to add as the unique identifier}

\item{col_name}{quoted string to use for the ID column. For example,
\code{tin_tare_set}}
}
\value{
a data frame with the ID appended (i.e. a new column with the ID
value recycled as an entry for each row)
}
\description{
Helper function for generating lists of equipment with corresponding IDs and
column names
}
