% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/astm_152H_hydrometers.R
\docType{data}
\name{astm_152H_hydrometers}
\alias{astm_152H_hydrometers}
\title{Dimensions and metadata on ASTM 152H hydrometers}
\format{
a tibble with one row per apparatus. Each column contains either an identifying feature of the hydrometer or a dimension.
}
\usage{
astm_152H_hydrometers
}
\description{
Used for computing effective depth, etc. during particle size analysis
}
\seealso{
\href{https://www.astm.org/Standards/D7928.htm}{ASTM D7928-17}
}
\keyword{datasets}
