% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bouyoucos_cylinders.R
\docType{data}
\name{bouyoucos_cylinders}
\alias{bouyoucos_cylinders}
\title{Dimensions and volumes of sedimentation cylinders used for particle size analysis}
\format{
Tibble with one row per cylinder.
}
\usage{
bouyoucos_cylinders
}
\description{
Used primarily for hydrometer analysis because cross-sectional area is required to compute effective depth and \% passing.
}
\keyword{datasets}
