% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowry_pipettes.R
\docType{data}
\name{lowry_pipettes}
\alias{lowry_pipettes}
\title{A set of lowry pipettes used for particle size analysis and making stock solutions}
\format{
A tibble with three columns: the set ID (a date formatted as a character string), the pipette number, and the volume in cm^3^
}
\usage{
lowry_pipettes
}
\description{
Precise volume as determined by gravimetric calibration with de-aired water permits precise correction for particle size analysis and stock solution concentrations.
}
\keyword{datasets}
