% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mini_sand_cone_vol.R
\docType{data}
\name{mini_sand_cone_vol}
\alias{mini_sand_cone_vol}
\title{Volume of Small-sample Bulk Density Tool}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
mini_sand_cone_vol
}
\description{
use for in-situ bulk density measurements on cleat mark cylinders.
}
\references{
\href{https://www.astm.org/Standards/D1556.htm}{ASTM D1556-15e1}
}
\keyword{datasets}
