% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proctor_cylinders.R
\docType{data}
\name{proctor_cylinders}
\alias{proctor_cylinders}
\title{A set of 4" diameter Proctor compaction cylinders}
\format{
A tibble with three columns: the cylinder number, the mass in grams, and the volume in cm^3^
}
\usage{
proctor_cylinders
}
\description{
Mass and volume data for each cylinder allows precise measurements for dry
density.
}
\keyword{datasets}
