% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proctor_molds.R
\docType{data}
\name{proctor_molds}
\alias{proctor_molds}
\title{A set of 4" diameter Proctor compaction molds}
\format{
A tibble with three columns: the mold ID, the mass in grams, and the volume in cm^3^
}
\usage{
proctor_molds
}
\description{
Mass and volume data for each cylinder allows precise measurements for dry
density.
}
\keyword{datasets}
