% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psa_beaker_tares.R
\docType{data}
\name{psa_beaker_tares}
\alias{psa_beaker_tares}
\title{Tared beaker masses}
\format{
a list of tibbles, with each containing two columns (the beaker number and the tared mass in grams, to 1 mg)
}
\usage{
psa_beaker_tares
}
\description{
These beakers are used repeatedly for particle size analysis and other lab work. Storing them in a list permits easy left joins and allows for addition of future expansion or new sets of beakers to be added.
}
\keyword{datasets}
