% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvc-perc-cylinders.R
\docType{data}
\name{pvc_perc_cylinders}
\alias{pvc_perc_cylinders}
\title{2" diameter PVC pipe sections}
\format{
A tibble containing the heights, masses, and cross-sectional areas of the cylinders.
}
\usage{
pvc_perc_cylinders
}
\description{
Used for K~sat~ and bulk density tests
}
\keyword{datasets}
