% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stainless_bowl_tares.R
\docType{data}
\name{stainless_bowl_tares}
\alias{stainless_bowl_tares}
\title{Stainless steel bowl masses for weighing moist soil}
\format{
Data frame containing 3 columns: the date (as a character type),
the number, and the mass to +/- 0.01 g.
}
\usage{
stainless_bowl_tares
}
\description{
Use for cleat mark testing, particle size analysis, etc.
}
\keyword{datasets}
