% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tin_tares.R
\docType{data}
\name{tin_tares}
\alias{tin_tares}
\title{Masses of small water content tins}
\format{
Data frame containing three columns: a character string which is the identifier for the tin tare set (typically a date in YYYY-MM-DD format),the tin number, and its mass in g.
}
\usage{
tin_tares
}
\description{
Tared masses to nearest 0.001 g. Useful for Atterberg limits and
other small-sample water content determinations.
}
\keyword{datasets}
