% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volumetric_flasks.R
\docType{data}
\name{volumetric_flasks}
\alias{volumetric_flasks}
\title{A set of volumetric flasks used for specific gravity tests}
\format{
A tibble with four columns: the set ID (a date formatted as a character string), the bottle number, the mass in grams, and the volume in cm^3^
}
\usage{
volumetric_flasks
}
\description{
Mass and volume data for each flask allows precise measurements for specific
gravity of soil particles or making stock solutions.
}
\keyword{datasets}
