% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrc_sand_loose_density.R
\docType{data}
\name{vrc_sand_loose_density}
\alias{vrc_sand_loose_density}
\title{Loose dry bulk density of Valentine topdressing sand}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
vrc_sand_loose_density
}
\description{
use for in-situ bulk density measurements on cleat mark cylinders and
backfilling cleat indentations to compute their volume.
}
\references{
\href{https://www.astm.org/Standards/D1556.htm}{ASTM D1556-15e1}
}
\keyword{datasets}
