% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psa_beaker_tares.R
\docType{data}
\name{psa_beaker_tares}
\alias{psa_beaker_tares}
\title{Tared beaker masses}
\format{
List of data frames, with each data frame containing three columns: an identifier for the set (typically a date), the beaker number and the tared mass in grams.
}
\usage{
psa_beaker_tares
}
\description{
\loadmathjax These beakers are used repeatedly for particle size
analysis and other lab work. Storing them in a list permits easy left joins
and allows for addition of future expansion or new sets of beakers to be
added. Mass is recorded to \mjeqn{\pm}{} 0.001 g)
}
\keyword{datasets}
