% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tin_tares.R
\docType{data}
\name{tin_tares}
\alias{tin_tares}
\title{Masses of water content tins}
\format{
a list containing a tibble for each set of tins. The tibble for a
given set contains only two columns: the tin number and its mass in g.
}
\usage{
tin_tares
}
\description{
Tared masses to nearest 0.001 g. Useful for Atterberg limits and
other small-sample water content determinations.
}
\keyword{datasets}
